# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/workforce/workers": {
      "get": {
        "operationId": "WorkforceController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workforce"
        ]
      },
      "post": {
        "operationId": "WorkforceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkerProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workforce"
        ]
      }
    },
    "/workforce/workers/me": {
      "get": {
        "operationId": "WorkforceController_findOwn",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workforce"
        ]
      }
    },
    "/workforce/workers/{id}": {
      "get": {
        "operationId": "WorkforceController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workforce"
        ]
      },
      "patch": {
        "operationId": "WorkforceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workforce"
        ]
      },
      "delete": {
        "operationId": "WorkforceController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workforce"
        ]
      }
    },
...
    "/complaints": {
      "get": {
        "operationId": "ComplaintsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      },
      "post": {
        "operationId": "ComplaintsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityComplaintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}": {
      "get": {
        "operationId": "ComplaintsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      },
      "patch": {
        "operationId": "ComplaintsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityComplaintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      },
      "delete": {
        "operationId": "ComplaintsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}/status": {
      "patch": {
        "operationId": "ComplaintsController_transitionStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionComplaintStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}/eligible-workers": {
      "get": {
        "operationId": "ComplaintsController_eligibleWorkers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}/assign": {
      "post": {
        "operationId": "ComplaintsController_assignWorker",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignWorkerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}/start": {
      "patch": {
        "operationId": "ComplaintsController_startWork",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}/resolve": {
      "patch": {
        "operationId": "ComplaintsController_resolveWork",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
    "/complaints/{id}/review": {
      "post": {
        "operationId": "ComplaintsController_submitReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      },
      "get": {
        "operationId": "ComplaintsController_findReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaints"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urbanity API",
    "description": "In-memory NestJS CRUD API with JWT-backed role-based access control.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Asha Rao",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "example": "asha@example.com",
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "example": "a-strong-development-password",
            "minLength": 8,
            "maxLength": 128
          },
          "phone": {
            "type": "string",
            "example": "+91-9876543210",
            "maxLength": 20
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "COMMUNITY_ADMIN",
              "TOWER_REPRESENTATIVE",
              "RESIDENT",
              "MAINTENANCE_WORKER"
            ],
            "example": "RESIDENT"
          },
          "communityId": {
            "type": "string",
            "example": "7b9f22cf-bbd3-4cc9-8d75-f21f02835b74"
          },
          "apartmentId": {
            "type": "string",
            "example": "7b9f22cf-bbd3-4cc9-8d75-f21f02835b74"
          },
          "towerId": {
            "type": "string",
            "example": "7b9f22cf-bbd3-4cc9-8d75-f21f02835b74"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Asha Rao",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "example": "asha@example.com",
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "example": "a-strong-development-password",
            "minLength": 8,
            "maxLength": 128
          },
          "phone": {
            "type": "string",
            "example": "+91-9876543210",
            "maxLength": 20
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "COMMUNITY_ADMIN",
              "TOWER_REPRESENTATIVE",
              "RESIDENT",
              "MAINTENANCE_WORKER"
            ],
            "example": "RESIDENT"
          },
          "communityId": {
            "type": "string",
            "example": "7b9f22cf-bbd3-4cc9-8d75-f21f02835b74"
          },
          "apartmentId": {
            "type": "string",
            "example": "7b9f22cf-bbd3-4cc9-8d75-f21f02835b74"
          },
          "towerId": {
            "type": "string",
            "example": "7b9f22cf-bbd3-4cc9-8d75-f21f02835b74"
          }
        }
      },
...
      "CreateWorkerProfileDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "specialization": {
            "type": "string",
            "enum": [
              "PLUMBING",
              "ELECTRICAL",
              "CARPENTRY",
              "HVAC",
              "LIFT_MAINTENANCE",
              "CLEANING",
              "GENERAL_MAINTENANCE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "BUSY",
              "ON_LEAVE",
              "INACTIVE"
            ],
            "default": "AVAILABLE"
          }
        },
        "required": [
          "userId",
          "specialization"
        ]
      },
      "UpdateWorkerProfileDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "specialization": {
            "type": "string",
            "enum": [
              "PLUMBING",
              "ELECTRICAL",
              "CARPENTRY",
              "HVAC",
              "LIFT_MAINTENANCE",
              "CLEANING",
              "GENERAL_MAINTENANCE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "BUSY",
              "ON_LEAVE",
              "INACTIVE"
            ],
            "default": "AVAILABLE"
          }
        }
      },
...
      "CreateCommunityComplaintDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "TOWER",
              "COMMUNITY"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "requiredWorkType": {
            "type": "string",
            "enum": [
              "PLUMBING",
              "ELECTRICAL",
              "CARPENTRY",
              "HVAC",
              "LIFT_MAINTENANCE",
              "CLEANING",
              "GENERAL_MAINTENANCE"
            ]
          },
          "apartmentId": {
            "type": "string"
          },
          "towerId": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "description",
          "requiredWorkType"
        ]
      },
      "UpdateCommunityComplaintDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "TransitionComplaintStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "UNDER_REVIEW",
              "ASSIGNED",
              "IN_PROGRESS",
              "RESOLVED",
              "REVIEWED",
              "CLOSED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "AssignWorkerDto": {
        "type": "object",
        "properties": {
          "workerId": {
            "type": "string"
          }
        },
        "required": [
          "workerId"
        ]
      },
      "WorkerActionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateComplaintReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "feedback": {
            "type": "string"
          }
        },
        "required": [
          "rating"
        ]
      }
    }
  }
}